Extract OpenAPI spec→tool machinery into src/openapi.ts
server.ts had grown to ~650 lines mixing two concerns: MCP server/tool
wiring and the OpenAPI-spec-to-tool mapping. Move the spec machinery
(pathToToolName, buildInputSchema, the OperationInfo shape, and the
SPEC_TYPES declaration inlined into the search tool) into a dedicated
src/openapi.ts. server.ts now imports them and is back to ~500 lines
focused on registering tools. No behavior change; tests updated to import
from the new module.
